restructure model
