+++ dbsetup.js
@@ -1,12 +1,8 @@
-const Account = require("./models/account");
-const Admin = require("./models/admin");
-const AdminGroup = require("./models/admin-group");
-const AuthAttempt = require("./models/auth-attempt");
 const MongoModels = require("mongo-models");
 const Promptly = require("promptly");
-const Session = require("./models/session");
-const Status = require("./models/status");
-const User = require("./models/user");
+
+const { Account, Admin, AdminGroup, Status } = require("./models/userRoles");
+const { AuthAttempt, Session, User } = require("./models/user");
 
 const main = async function() {
     let options = {};
