+++ graphql/type/post/post.graphql
@@ -1,0 +1,13 @@
+# import Node from "../interface.graphql"
+# import Image from "../general/util.graphql"
+
+### Post will contain a header image, title, description and maybe extra image
+type Post implements Node {
+    id: ID!
+    createdBy: User!
+    createdTime: String!
+    title: String!
+    description: [String]!
+    headerImage: Image!
+    extraImages: [Image]
+}
