+++ models/user/index.js
@@ -1,0 +1,9 @@
+const User = require("./user");
+const AuthAttempt = require("./auth-attempt");
+const Session = require("./session");
+
+module.exports = {
+    User,
+    AuthAttempt,
+    Session,
+};
