+++ models/userRoles/index.js
@@ -1,0 +1,15 @@
+const Admin = require("./admin");
+const AdminGroup = require("./admin-group");
+const Account = require("./account");
+const NoteEntry = require("./note-entry");
+const StatusEntry = require("./status-entry");
+const Status = require("./status");
+
+module.exports = {
+    Admin,
+    AdminGroup,
+    Account,
+    NoteEntry,
+    Status,
+    StatusEntry,
+};
